+++ cmd/action/verify/kiam/awsapicall/runner.go
@@ -10,8 +10,7 @@ import (
 	"github.com/spf13/cobra"
 	batchapiv1 "k8s.io/api/batch/v1"
 	apiv1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
+	k8sruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/awscnfm/v12/pkg/client"
 	"github.com/giantswarm/awscnfm/v12/pkg/env"
@@ -75,7 +74,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		}
 	}
 
-	err = r.checkKiamAWScallPod(ctx, tcClients.K8sClient())
+	err = r.checkKiamAWScallPod(ctx, tcClients.CtrlClient())
 	if err != nil {
 		return microerror.Mask(err)
 	}
@@ -84,11 +83,19 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 }
 
 // checkKiamAWScallPod will check if the AWS API call job was successful
-func (r *runner) checkKiamAWScallPod(ctx context.Context, tcClient kubernetes.Interface) error {
-
+func (r *runner) checkKiamAWScallPod(ctx context.Context, tcClient k8sruntimeclient.Client) error {
 	name := fmt.Sprintf("%s-kiam-test", project.Name())
 
-	job, err := tcClient.BatchV1().Jobs(kubeSystemNamespace).Get(ctx, name, metav1.GetOptions{})
+	job := &batchapiv1.Job{}
+
+	err := tcClient.Get(
+		ctx,
+		k8sruntimeclient.ObjectKey{
+			Namespace: kubeSystemNamespace,
+			Name:      name,
+		},
+		job,
+	)
 	if err != nil {
 		return microerror.Mask(err)
 	}